+++ lib/components/index.ts
@@ -1,3 +1,5 @@
+import Portal from '@/components/Portal'
+import UseClickOutside from '@/components/UseClickOutside'
 import UseSwipe from '@/components/UseSwipe'
 
-export { UseSwipe }
+export { UseSwipe, Portal, UseClickOutside }